AAP-884 Flytt markering-typer til kontrakt for frontend
