+++ api-kontrakt/src/main/kotlin/no/nav/aap/oppgave/markering/MarkeringDto.kt
@@ -1,6 +1,6 @@
 package no.nav.aap.oppgave.markering
 
-import no.nav.aap.oppgave.MarkeringForBehandling
+import no.nav.aap.oppgave.verdityper.MarkeringForBehandling
 
 data class MarkeringDto(
     val type: MarkeringForBehandling,
