+++ api-kontrakt/src/main/kotlin/no/nav/aap/oppgave/markering/MarkeringResponse.kt
@@ -1,6 +1,6 @@
 package no.nav.aap.oppgave.markering
 
-import no.nav.aap.oppgave.MarkeringForBehandling
+import no.nav.aap.oppgave.verdityper.MarkeringForBehandling
 
 data class MarkeringResponse(
     val markeringType: MarkeringForBehandling,
