+++ api-kontrakt/src/main/kotlin/no/nav/aap/oppgave/verdityper/MarkeringForBehandling.kt
@@ -1,0 +1,6 @@
+package no.nav.aap.oppgave.verdityper
+
+enum class MarkeringForBehandling {
+    HASTER,
+    KREVER_SPESIALKOMPETANSE
+}
