+++ app/src/main/kotlin/no/nav/aap/oppgave/oppgaveliste/OppgavelisteService.kt
@@ -4,7 +4,6 @@ import no.nav.aap.komponenter.httpklient.httpclient.tokenprovider.OidcToken
 import no.nav.aap.komponenter.miljo.Miljø
 import no.nav.aap.komponenter.miljo.MiljøKode
 import no.nav.aap.oppgave.AvklaringsbehovReferanseDto
-import no.nav.aap.oppgave.BehandlingMarkering
 import no.nav.aap.oppgave.OppgaveDto
 import no.nav.aap.oppgave.OppgaveRepository
 import no.nav.aap.oppgave.OppgaveRepository.FinnOppgaverDto
@@ -12,7 +11,9 @@ import no.nav.aap.oppgave.enhet.EnhetService
 import no.nav.aap.oppgave.filter.FilterDto
 import no.nav.aap.oppgave.liste.Paging
 import no.nav.aap.oppgave.liste.UtvidetOppgavelisteFilter
+import no.nav.aap.oppgave.markering.BehandlingMarkering
 import no.nav.aap.oppgave.markering.MarkeringRepository
+import no.nav.aap.oppgave.markering.MarkeringResponse
 import no.nav.aap.oppgave.unleash.FeatureToggles
 import no.nav.aap.oppgave.unleash.IUnleashService
 import no.nav.aap.oppgave.unleash.UnleashServiceProvider
@@ -36,7 +37,7 @@ class OppgavelisteService(
                 "Fant ikke behandlingsreferanse for oppgave med id ${oppgave.id}"
             }
             val markeringer = markeringRepository.hentMarkeringerForBehandling(behandlingRef)
-            oppgave.leggPåMarkeringer(markeringer)
+            oppgave.leggPåMarkeringer(markeringer.map { it.tilMarkeringResponse() })
         }
     }
 
@@ -44,7 +45,7 @@ class OppgavelisteService(
         val oppgave = oppgaveRepository.hentOppgave(avklaringsbehovReferanseDto)
         if (avklaringsbehovReferanseDto.referanse != null) {
             val markeringer = markeringRepository.hentMarkeringerForBehandling(avklaringsbehovReferanseDto.referanse!!)
-            return oppgave?.leggPåMarkeringer(markeringer)
+            return oppgave?.leggPåMarkeringer(markeringer.map { it.tilMarkeringResponse() })
         }
         return oppgave
     }
@@ -93,7 +94,7 @@ class OppgavelisteService(
                     "Fant ikke behandlingsreferanse for oppgave med id ${oppgave.id}"
                 }
                 val markeringer = markeringRepository.hentMarkeringerForBehandling(behandlingRef)
-                oppgave.leggPåMarkeringer(markeringer)
+                oppgave.leggPåMarkeringer(markeringer.map { it.tilMarkeringResponse() })
             }
 
         return FinnOppgaverDto(
@@ -111,7 +112,7 @@ class OppgavelisteService(
             it.leggPåMarkeringer(
                 markeringRepository.hentMarkeringerForBehandling(requireNotNull(it.behandlingRef) {
                     "Fant ikke behandlingsreferanse for oppgave med id ${it.id}"
-                })
+                }).map { it.tilMarkeringResponse() }
             )
         }.medPersonNavn(fjernSensitivInformasjonNårTilgangMangler = false, token = token)
 
@@ -126,7 +127,7 @@ class OppgavelisteService(
         )
     }
 
-    private fun OppgaveDto.leggPåMarkeringer(markeringer: List<BehandlingMarkering>): OppgaveDto =
+    private fun OppgaveDto.leggPåMarkeringer(markeringer: List<MarkeringResponse>): OppgaveDto =
         this.copy(markeringer = markeringer)
 
     private fun List<OppgaveDto>.filtrerPåTilgang(
@@ -153,4 +154,12 @@ class OppgavelisteService(
         } else {
             oppgaver
         }
+}
+
+fun BehandlingMarkering.tilMarkeringResponse(): MarkeringResponse {
+    return MarkeringResponse(
+        this.markeringType,
+        this.begrunnelse,
+        this.opprettetAv
+    )
 }
